usersの新規投稿のボクサー選択と新規ボクサー作成時のジム選択を検索できるセレクトボックスに変更,header選択時の文字の色を修正,users周りの自分以外の投稿やコメントを編集・削除できないように制御,新規ボクサー作成は管理者権限とする制御を作成,フォロワーの名前をポップで表示,「投げ銭する」ボタン押下時に別タブで開くように修正,右上のプロフィールタブの色を背景色と同化させるように修正

+++ app/Http/Controllers/BoxerPostsController.php
@@ -6,6 +6,7 @@ use App\Models\Boxer;
 use App\Models\BoxerPost;
 use App\Http\Requests\StoreBoxerPostsRequest;
 use App\Http\Requests\UpdateBoxerPostsRequest;
+use Illuminate\Support\Facades\Auth;
 
 class BoxerPostsController extends Controller
 {
@@ -28,6 +29,8 @@ class BoxerPostsController extends Controller
     {
         $validated = $request->validated();
 
+        $validated['user_id'] = Auth::id();
+
         $boxer->posts()->create($validated);
 
         return redirect()->route('boxer_posts.index', $boxer);
@@ -35,11 +38,18 @@ class BoxerPostsController extends Controller
 
     public function edit(BoxerPost $post)
     {
+        if (Auth::id() !== $post->user_id) {
+            abort(403);
+        }
         return view('boxer_posts.edit', compact('post'));
     }
 
     public function update(UpdateBoxerPostsRequest $request, BoxerPost $post)
     {
+        if (Auth::id() !== $post->user_id) {
+            abort(403);
+        }
+
         $validated = $request->validated();
 
         $post->update($validated);
@@ -49,6 +59,10 @@ class BoxerPostsController extends Controller
 
     public function destroy(BoxerPost $post)
     {
+        if (Auth::id() !== $post->user_id) {
+            abort(403);
+        }
+
         $post->delete();
 
         return back();
